finally got bullet collision detection working without the whole tunneling problem
granted, multiple enemy penetration doesn't work as favorably as it
should; but that was more of a side-effect in ranger scratch.
between that and actually having accurate detection, i'll take the
latter.

I think i just won't do inter-entity collision resolution. It definitely
wasn't in the original, so i won't bother.

map collisions can be solved quite easily with a grid-based map; which
i'm starting to be partial to.