yeah no i give up
i'm not fucking around with collisions any longer. we have the basic
static resolution working, and that's more than enough for me.

the most important element, the bullet detection, works flawlessly now.
so i have absolutely ZERO reason to fuck around with this option feature
now.

will i pick it up again in the future? probably. is it going to be the
next commit? probably.