added grid field ids and assigning entities to fields by id
fixed a little bruh moment with the grid init function.
the only consequence was just that non-square grids don't fucking work
anymore. yippie

entities are to receive a component which holds all the fields it is
"assigned" to. this is probably best for bevy, because it can help us
prevent needing to recursively check wether an entity is within a given
grid field. i'd like to let the ecs do as much work as possible.