added a better version of the assignment to grid field thing
i guess assignment isn't quite the right term, association is better.
anyway; given a point or aabb the grid returns a collection of
row/column locations, which can be used to index into a field.

this can be used, for instance, when you wanna see if an actor is
permitted to enter into the next field; and if not, preventing it.

TODO:
- refactor `assign` into `associate`
- uncomment the field select through row/column method