did what i mentioned last time and fixed a little ecs error
since we have a sizeable reliance on marker components being swapped in
and out, it is vital to perform the insertions for this BEFORE removing
any entities. right now i have it done with explicit system ordering,
but it can get real tedious real fast to have to append each destructive
system with the required explicit ordering. as a result either
systemsets or labels will have to be implemented to make things
bearable.

https://bevy-cheatbook.github.io/programming/system-order.html
https://bevy-cheatbook.github.io/programming/system-sets.html