removed the collision resolution stuff again
It was really shit, so I'd rather start over with a proper understanding
of what the hell I'm doing. It also, y'know, didn't work.

replaced the horrific raycasting code and replaced it with a much more
simplified version.
collisions are obviously not detected perfectly, because a given enemy
is also moving most of the time; which hasn't been accounted for yet.